Added and employed a method for optionally retrieving values at paths.
The new method will return Optional<Value>.None when a JSON value is found at the specified key path, but that JSON value is .Null.  Employed new method in optionalAtPath(_:_:ifNotFound:ifNull:getter:) in the case that ifNull is true.  As a result, when ifNull is true, callers will get back nil from optionalAtPath(_:_:ifNotFound:ifNull:getter:) if a JSON value of .None is found at that path.