Refactor number parsing to detect over/underflow.
Pulls all the (private) number parsing methods out of JSONParser and
into (private) NumberParser. Instead of building up the numeric value as
it's parsed, NumberParser finds the end of the number, stuffs it into a
String, then attempts to convert it to an Int or Double (depending on
whether we saw decimal features during parsing). This allows us to catch
unrepresentable values and throw an error instead of crashing.