NSJSONSerialization: 32-bit overflow becomes Double
Prior to this, we would not crash when using the NSJSONSerialization
parser, but we would mangle the value by incorrectly assuming "not
a char and not a float means it fits in an Int", even when it was
properly an Int64.

We might want to force this overflow case to String for parity with the
Freddy parser.