Adds private method to decode terminating null into nil [Fixes #173]
Users of Freddy had trouble decoding JSON that had a terminating Null into some JSONDecodable type. This commit adds a private static method to JSON that checks to see if the JSON in question is equal to .Null. If it is, then Error.ValueNotConvertible is thrown. Thus, if .NullBecomesNil along the path in question, then mapOptionalAtPath(_:alongPath:transform:) will return nil.