Resolve crash when string ends in backslash
The problem was that the parser would read beyond a backslash even if it
was the last character, so rather than reporting it as invalid JSON and
throwing an error it would try to read beyond the string and crash. This
fix just adds a check to ensure that the parser doesn't read out of
bounds.